refactor(permission): replace RequireSuperAdmin with SuperAdminOnly option for enhanced role checks
