+++ internal/rest/middleware/permission.go
@@ -24,17 +24,66 @@ func NewPermissionMiddleware(rbacService *rbac.RBACService, logger *logger.Logge
 	}
 }
 
-// RequirePermission returns a middleware that enforces two access controls:
-// suspended tenants are blocked from write operations regardless of caller type,
-// and service accounts are subject to RBAC role checks for the given entity and action.
-func (pm *PermissionMiddleware) RequirePermission(entity types.Entity, action types.Action) gin.HandlerFunc {
+// PermissionOption declares an extra rule a route requires on top of its
+// entity/action pair. Options are resolved once when the route is registered.
+// Add an option here rather than a new middleware when a route needs a stricter
+// policy, for example a future route limited to a specific role.
+type PermissionOption func(*permissionRules)
+
+type permissionRules struct {
+	superAdminOnly bool
+}
+
+// SuperAdminOnly restricts a route to a person holding super_admin. Use it for
+// routes that administer people, such as changing another user's roles, where a
+// service account must be refused no matter which roles its key was minted with.
+//
+// The caller kind is tested with IsServiceAccount rather than by requiring
+// "user": a JWT session carries no user type at all (see setContextValues in
+// auth.go), so an equality check would reject every dashboard user.
+func SuperAdminOnly() PermissionOption {
+	return func(r *permissionRules) { r.superAdminOnly = true }
+}
+
+// RequirePermission gates a route on the caller's roles for the given entity and
+// action, plus any options the route opts into. Checks run in order of how
+// fundamental they are, so the caller is told about the problem it must fix
+// first: a suspended tenant, then the route's own rules, then plain RBAC.
+func (pm *PermissionMiddleware) RequirePermission(entity types.Entity, action types.Action, opts ...PermissionOption) gin.HandlerFunc {
+	var rules permissionRules
+	for _, opt := range opts {
+		opt(&rules)
+	}
+
 	return func(c *gin.Context) {
 		ctx := c.Request.Context()
 
-		// Suspended tenants are blocked from all write operations.
+		// Suspended tenants are blocked from all write operations. Checked first
+		// so a suspended tenant hears about the suspension rather than a narrower
+		// role problem it cannot act on until the account is restored.
 		if action == types.ActionWrite && types.GetTenantInternalStatus(ctx) == types.TenantInternalStatusSuspended {
 			c.AbortWithStatusJSON(http.StatusForbidden, gin.H{
 				"message": "tenant account is suspended",
+			})
+			return
+		}
+
+		if rules.superAdminOnly && (types.IsServiceAccount(ctx) || !lo.Contains(types.GetRoles(ctx), types.RoleSuperAdmin.String())) {
+			pm.logger.Info(ctx, "access denied: administrative action requires a super_admin user",
+				"user_id", types.GetUserID(ctx),
+				"tenant_id", types.GetTenantID(ctx),
+				"environment_id", types.GetEnvironmentID(ctx),
+				"roles", types.GetRoles(ctx),
+				"is_service_account", types.IsServiceAccount(ctx),
+				"entity", entity,
+				"action", action,
+				"path", c.Request.URL.Path,
+			)
+			// Names both requirements, so it reads correctly whichever half fails:
+			// a service account fails on "user account" even when its key carries
+			// super_admin, and a plain user fails on "Super Admin access".
+			c.AbortWithStatusJSON(http.StatusForbidden, gin.H{
+				"message": "This action requires a user account with Super Admin access",
 			})
 			return
 		}
@@ -69,45 +118,5 @@ func (pm *PermissionMiddleware) RequirePermission(entity types.Entity, action ty
 		}
 
 		c.Next()
-	}
-}
-
-// RequireSuperAdmin gates administrative routes such as changing another user's
-// roles. On top of the usual entity/action check it demands that the caller is a
-// person holding super_admin, so a service account can never administer the
-// people in a tenant no matter which roles its key was minted with.
-//
-// The caller kind is tested with IsServiceAccount rather than by requiring
-// "user": a JWT session carries no user type at all (see setContextValues in
-// auth.go), so an equality check would reject every dashboard user.
-func (pm *PermissionMiddleware) RequireSuperAdmin(entity types.Entity, action types.Action) gin.HandlerFunc {
-	return func(c *gin.Context) {
-		ctx := c.Request.Context()
-
-		if types.IsServiceAccount(ctx) || !lo.Contains(types.GetRoles(ctx), types.RoleSuperAdmin.String()) {
-			pm.logger.Info(ctx, "access denied: administrative action requires a super_admin user",
-				"user_id", types.GetUserID(ctx),
-				"tenant_id", types.GetTenantID(ctx),
-				"environment_id", types.GetEnvironmentID(ctx),
-				"roles", types.GetRoles(ctx),
-				"is_service_account", types.IsServiceAccount(ctx),
-				"entity", entity,
-				"action", action,
-				"path", c.Request.URL.Path,
-			)
-			// Names both requirements, so it reads correctly for either half of
-			// the guard: a service account fails on "user account" even when its
-			// key carries super_admin, and a plain user fails on "Super Admin
-			// access". Deliberately not "insufficient permissions", which
-			// RequirePermission already returns for an ordinary RBAC miss.
-			c.AbortWithStatusJSON(http.StatusForbidden, gin.H{
-				"message": "This action requires a user account with Super Admin access",
-			})
-			return
-		}
-
-		// Falls through to the standard check so suspended tenants and the RBAC
-		// rules stay enforced on this route too.
-		pm.RequirePermission(entity, action)(c)
 	}
 }
