+++ internal/rest/middleware/permission_test.go
@@ -55,6 +55,180 @@ func newPermissionTestRouter(t *testing.T, rbacSvc *rbac.RBACService, userType s
 	return r
 }
 
+// superAdminRolesJSON gives all_writer real write permission on every entity, so
+// a caller rejected under SuperAdminOnly is provably rejected by that option
+// rather than by the underlying RBAC check.
+const superAdminRolesJSON = `{
+	"super_admin": {
+		"name": "Super Admin",
+		"permissions": { "*": ["*"] }
+	},
+	"all_writer": {
+		"name": "All Writer",
+		"permissions": { "*": ["read", "write"] }
+	},
+	"event_ingestor": {
+		"name": "Event Ingestor",
+		"permissions": { "event": ["write"] }
+	}
+}`
+
+// newSuperAdminRouter gates PUT /users/:id/roles with the SuperAdminOnly option
+// and runs it behind TenantStatusMiddleware, mirroring the real route.
+func newSuperAdminRouter(t *testing.T, tenantStatus types.TenantInternalStatus, userType string, roles []string) *gin.Engine {
+	t.Helper()
+	log := newTestLogger(t)
+	pm := NewPermissionMiddleware(newRBACServiceWithRoles(t, superAdminRolesJSON), log)
+
+	gin.SetMode(gin.TestMode)
+	r := gin.New()
+
+	r.Use(func(c *gin.Context) {
+		ctx := c.Request.Context()
+		ctx = context.WithValue(ctx, types.CtxTenantID, "tenant-rbac")
+		ctx = context.WithValue(ctx, types.CtxUserID, "user-1")
+		if userType != "" {
+			ctx = context.WithValue(ctx, types.CtxUserType, userType)
+		}
+		if roles != nil {
+			ctx = context.WithValue(ctx, types.CtxRoles, roles)
+		}
+		c.Request = c.Request.WithContext(ctx)
+		c.Next()
+	})
+
+	r.Use(TenantStatusMiddleware(&mockTenantService{status: tenantStatus}, log))
+	r.PUT("/users/:id/roles",
+		pm.RequirePermission(types.EntityUser, types.ActionWrite, SuperAdminOnly()),
+		func(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"ok": true}) },
+	)
+
+	return r
+}
+
+func TestRequirePermission_SuperAdminOnly(t *testing.T) {
+	const wantAdminMsg = "This action requires a user account with Super Admin access"
+
+	testCases := []struct {
+		name         string
+		userType     string
+		roles        []string
+		tenantStatus types.TenantInternalStatus
+		wantCode     int
+		wantErrMsg   string
+	}{
+		{
+			// A dashboard session carries no user type at all, so this is the
+			// ordinary success path and would break under a userType == "user" check.
+			name:         "super_admin JWT user with no user type is allowed",
+			userType:     "",
+			roles:        []string{types.RoleSuperAdmin.String()},
+			tenantStatus: types.TenantInternalStatusActive,
+			wantCode:     http.StatusOK,
+		},
+		{
+			name:         "super_admin user is allowed",
+			userType:     string(types.UserTypeUser),
+			roles:        []string{types.RoleSuperAdmin.String()},
+			tenantStatus: types.TenantInternalStatusActive,
+			wantCode:     http.StatusOK,
+		},
+		{
+			// all_writer passes the entity/action check, so only SuperAdminOnly
+			// can be what refuses it here.
+			name:         "user with write access but not super_admin is denied",
+			userType:     string(types.UserTypeUser),
+			roles:        []string{types.RoleAllWriter.String()},
+			tenantStatus: types.TenantInternalStatusActive,
+			wantCode:     http.StatusForbidden,
+			wantErrMsg:   wantAdminMsg,
+		},
+		{
+			// The key carries super_admin, yet administering people stays closed
+			// to machine credentials.
+			name:         "service account holding super_admin is denied",
+			userType:     string(types.UserTypeServiceAccount),
+			roles:        []string{types.RoleSuperAdmin.String()},
+			tenantStatus: types.TenantInternalStatusActive,
+			wantCode:     http.StatusForbidden,
+			wantErrMsg:   wantAdminMsg,
+		},
+		{
+			name:         "service account without super_admin is denied",
+			userType:     string(types.UserTypeServiceAccount),
+			roles:        []string{types.RoleEventIngestor.String()},
+			tenantStatus: types.TenantInternalStatusActive,
+			wantCode:     http.StatusForbidden,
+			wantErrMsg:   wantAdminMsg,
+		},
+		{
+			name:         "caller with no roles is denied",
+			userType:     string(types.UserTypeUser),
+			roles:        nil,
+			tenantStatus: types.TenantInternalStatusActive,
+			wantCode:     http.StatusForbidden,
+			wantErrMsg:   wantAdminMsg,
+		},
+		{
+			name:         "caller with an empty role set is denied",
+			userType:     string(types.UserTypeUser),
+			roles:        []string{},
+			tenantStatus: types.TenantInternalStatusActive,
+			wantCode:     http.StatusForbidden,
+			wantErrMsg:   wantAdminMsg,
+		},
+		{
+			name:         "super_admin in a suspended tenant is denied",
+			userType:     string(types.UserTypeUser),
+			roles:        []string{types.RoleSuperAdmin.String()},
+			tenantStatus: types.TenantInternalStatusSuspended,
+			wantCode:     http.StatusForbidden,
+			wantErrMsg:   "tenant account is suspended",
+		},
+		{
+			// Suspension is the more fundamental problem and nothing the caller
+			// can fix with a role change, so it must win over the admin guard.
+			name:         "suspension is reported ahead of the super_admin requirement",
+			userType:     string(types.UserTypeUser),
+			roles:        []string{types.RoleAllWriter.String()},
+			tenantStatus: types.TenantInternalStatusSuspended,
+			wantCode:     http.StatusForbidden,
+			wantErrMsg:   "tenant account is suspended",
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			router := newSuperAdminRouter(t, tc.tenantStatus, tc.userType, tc.roles)
+
+			w := httptest.NewRecorder()
+			req, _ := http.NewRequest(http.MethodPut, "/users/user-2/roles", nil)
+			router.ServeHTTP(w, req)
+
+			assert.Equal(t, tc.wantCode, w.Code)
+
+			if tc.wantErrMsg != "" {
+				var body map[string]string
+				require.NoError(t, json.Unmarshal(w.Body.Bytes(), &body))
+				assert.Equal(t, tc.wantErrMsg, body["message"])
+			}
+		})
+	}
+}
+
+// Without the super_admin guard the same route admits any caller with entity write access,
+// which is the escalation SuperAdminOnly exists to close.
+func TestRequirePermission_PlainCheckAllowsWriter(t *testing.T) {
+	router := newPermissionTestRouter(t, newRBACServiceWithRoles(t, superAdminRolesJSON),
+		string(types.UserTypeUser), []string{types.RoleAllWriter.String()},
+		types.EntityUser, types.ActionWrite)
+
+	w := httptest.NewRecorder()
+	router.ServeHTTP(w, httptest.NewRequest(http.MethodPost, "/customers", nil))
+
+	assert.Equal(t, http.StatusOK, w.Code, "all_writer passes a plain entity/action check")
+}
+
 // TestRequirePermission_DeniesServiceAccountWithoutRole reproduces the
 // live-verified staging PoC: a service-account key scoped to
 // event_ingestor/event_reader (neither of which grants customer:write)
